requirements.txt
albucore==0.0.12
albumentations==1.4.10
asttokens==2.4.1
attrs==23.2.0
Automat==22.10.0
blinker==1.8.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
constantly==23.10.4
cryptography==42.0.8
cssselect==1.2.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.15.4
Flask==3.0.3
hyperlink==21.0.0
idna==3.7
importlib_metadata==8.0.0
incremental==22.10.0
ipykernel==6.29.5
ipython==8.26.0
itemadapter==0.9.0
itemloaders==1.3.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
jupyter_client==8.6.2
jupyter_core==5.7.2
lazy_loader==0.4
lxml==5.2.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
nest_asyncio==1.6.0
packaging==24.1
parsel==1.9.1
parso==0.8.4
pickleshare==0.7.5
pip==24.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
Protego==0.3.1
psutil==5.9.0
psycopg2-binary==2.9.9
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.4
pydantic_core
PyDispatcher==2.0.7
Pygments==2.18.0
pyOpenSSL==24.1.0
python-dateutil==2.9.0
pyzmq==25.1.2
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
scikit-image==0.24.0
Scrapy==2.11.2
service-identity==24.1.0
setuptools==69.5.1
six==1.16.0
stack-data==0.6.2
tldextract==5.1.2
tornado==6.4.1
traitlets==5.14.3
Twisted==24.3.0
typing_extensions==4.12.2
urllib3==2.2.2
w3lib==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.3
wheel==0.43.0
zipp==3.19.2
zope.interface==6.4.post2
python-dotenv
gunicorn
flask-cors